backend/elasticsearch_model/app/requirements.txt
aniso8601==8.1.0
click==7.1.2
Cython==0.29.14
Flask==1.1.2
Flask-RESTful==0.3.8
gensim==3.8.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
scipy==1.5.4
six==1.15.0
smart-open==4.1.0
tqdm==4.56.0
Werkzeug==1.0.1
unidecode
tika
scikit-learn
python-docx
elasticsearch
paho-mqtt
